backend/requirements.txt
# backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.1
openai==1.14.0
reportlab==4.2.0
pydantic==2.7.1
python-multipart==0.0.9
httpx==0.27.0